feat: api 연동 및 다이나믹 라우팅 적용
